Update
The core of Crop Model MKS is reconstructed by C++. The COM is no longer used but the dynamic library directly. The system is thus not compatible to other programming language any more. I think the compatibility is useless because during the attempts to migrate other components from other crop models, I found to effort to archive a successful migration is no less than create a new one.